requirements.txt
Flask==3.1.0
openai==1.58.1
python-dotenv==1.0.1
gunicorn==21.2.0
